Harden register allocation correctness and tests
Fix CFG/phi-aware liveness and interference, validate SSA and assignments, harden code generation and native execution, add deterministic allocators and adversarial differential test coverage.

+++ crates/codegen/src/simd.rs
@@ -19,7 +19,7 @@ pub fn emit_sum8_scalar() -> Result<Vec<u8>> {
     a.xorps(xmm0, xmm0)?;
     for i in 0..8 {
         // movss xmm1, [rdi + i*4]
-        a.movss(xmm1, dword_ptr(rdi + (i*4) as i32))?;
+        a.movss(xmm1, dword_ptr(rdi + i * 4))?;
         a.addss(xmm0, xmm1)?;
     }
     // epilogue
